requirements/backend.txt
uvicorn
fastapi-offline #for testing purposes
plyer
pydantic
pydantic-settings
fastapi
boto3   #AWS file upload and referencing
PyJWT
httpx
pyinstrument
email-validator
aiofiles
redis
aioredis
sqlalchemy
fastapi-limiter
sqladmin
passlib
asyncpg
alembic
pyotp